feat: adds use case to get weather from a location

+++ data/src/main/java/com/lfaiska/weather/data/remote/dto/ConsolidatedWeatherResponse.kt
@@ -15,5 +15,5 @@ data class ConsolidatedWeatherResponse (
     @SerializedName("wind_speed")
     val windSpeed: Double,
     @SerializedName("humidity")
-    val humidity: Double
+    val humidity: Int
 )
